Error checking for length parameter of external array constructors in shell
BUG=v8:1501

Review URL: http://codereview.chromium.org/7268002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@8458 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/d8.h
@@ -217,7 +217,7 @@ class Shell: public i::AllStatic {
   static Counter* GetCounter(const char* name, bool is_histogram);
   static Handle<Value> CreateExternalArray(const Arguments& args,
                                            ExternalArrayType type,
-                                           int element_size);
+                                           size_t element_size);
   static void ExternalArrayWeakCallback(Persistent<Value> object, void* data);
 };
